Split move_program_factory.cpp into separate files\n\n- Extract move_multi_core_with_overlap and move_multi_core_sharded into their own .hpp/.cpp files\n- Update original file to include the new extracted files\n- Update CMakeLists.txt to compile the new source files\n- Remove extracted functions from the original file\n\nThis refactoring is part of the program factory splitting effort for ProgramDescriptor migration.
